Changed occurrences of uint34_t to size_t to prevent overflow and changed a dumb conversion function to a more suited one (that was an oversight i forgot to change begore)

+++ time_parser.h
@@ -3,8 +3,8 @@
 #ifndef TIME_PARSER
 #define TIME_PARSER
 
-uint32_t parse_time_ms(const char* str);
-uint32_t parse_time_sec(const char* str);
-uint32_t parse_time_min(const char* str);
+size_t parse_time_ms(const char* str);
+size_t parse_time_sec(const char* str);
+size_t parse_time_min(const char* str);
 
 #endif
